Refactor SectionBrowser to bottom of section
